added sitemap generation function
Integrates the `gulp-sitemap` plugin to enable the generation of a sitemap for the project. The new task, `generateSitemap`, creates a sitemap for search engines, enhancing SEO capabilities. Users can run `gulp sitemap/` to generate the sitemap.

This commit includes an updated README section providing instructions on using the new sitemap generation function.